Fixes for length

+++ podnoms-data/Models/PodcastEntry.cs
@@ -24,17 +24,14 @@ namespace PodNoms.Data.Models {
 
         public string Description { get; set; }
 
-        [MaxLength(2000)]
-        public string SourceUrl { get; set; }
+        [MaxLength(2000)] public string SourceUrl { get; set; }
 
-        [MaxLength(2000)]
-        public string AudioUrl { get; set; }
-        public float AudioLength { get; set; }
+        [MaxLength(2000)] public string AudioUrl { get; set; }
+        public double AudioLength { get; set; }
 
         public long AudioFileSize { get; set; }
 
-        [MaxLength(2000)]
-        public string ImageUrl { get; set; }
+        [MaxLength(2000)] public string ImageUrl { get; set; }
         public string ProcessingPayload { get; set; }
         public ProcessingStatus ProcessingStatus { get; set; } = ProcessingStatus.Accepted;
         public DateTime? SourceCreateDate { get; set; }
